+++ src/sstat/print.c
@@ -87,7 +87,7 @@ static void _print_small_double(
 		return;
 
 	if (dub > 1)
-		convert_num_unit((float)dub, outbuf, buf_size, units);
+		convert_num_unit((double)dub, outbuf, buf_size, units);
 	else if (dub > 0)
 		snprintf(outbuf, buf_size, "%.2fM", dub);
 	else
@@ -117,7 +117,7 @@ void print_fields(slurmdb_step_rec_t *step)
 			break;
 		case PRINT_ACT_CPUFREQ:
 
-			convert_num_unit2((float)step->stats.act_cpufreq,
+			convert_num_unit2((double)step->stats.act_cpufreq,
 					  outbuf, sizeof(outbuf),
 					  UNIT_KILO, 1000, false);
 
@@ -127,7 +127,7 @@ void print_fields(slurmdb_step_rec_t *step)
 			break;
 		case PRINT_CONSUMED_ENERGY:
 			if (!fuzzy_equal(step->stats.consumed_energy, NO_VAL)) {
-				convert_num_unit2((float)
+				convert_num_unit2((double)
 						  step->stats.consumed_energy,
 						  outbuf, sizeof(outbuf),
 						  UNIT_NONE, 1000, false);
@@ -160,7 +160,7 @@ void print_fields(slurmdb_step_rec_t *step)
 					     (curr_inx == field_count));
 			break;
 		case PRINT_AVEPAGES:
-			convert_num_unit((float)step->stats.pages_ave,
+			convert_num_unit((double)step->stats.pages_ave,
 					 outbuf, sizeof(outbuf),
 					 UNIT_KILO);
 
@@ -169,7 +169,7 @@ void print_fields(slurmdb_step_rec_t *step)
 					     (curr_inx == field_count));
 			break;
 		case PRINT_AVERSS:
-			convert_num_unit((float)step->stats.rss_ave,
+			convert_num_unit((double)step->stats.rss_ave,
 					 outbuf, sizeof(outbuf),
 					 UNIT_KILO);
 
@@ -178,7 +178,7 @@ void print_fields(slurmdb_step_rec_t *step)
 					     (curr_inx == field_count));
 			break;
 		case PRINT_AVEVSIZE:
-			convert_num_unit((float)step->stats.vsize_ave,
+			convert_num_unit((double)step->stats.vsize_ave,
 					 outbuf, sizeof(outbuf),
 					 UNIT_KILO);
 
@@ -246,7 +246,7 @@ void print_fields(slurmdb_step_rec_t *step)
 					     (curr_inx == field_count));
 			break;
 		case PRINT_MAXPAGES:
-			convert_num_unit((float)step->stats.pages_max,
+			convert_num_unit((double)step->stats.pages_max,
 					 outbuf, sizeof(outbuf),
 					 UNIT_KILO);
 
@@ -269,7 +269,7 @@ void print_fields(slurmdb_step_rec_t *step)
 					     (curr_inx == field_count));
 			break;
 		case PRINT_MAXRSS:
-			convert_num_unit((float)step->stats.rss_max,
+			convert_num_unit((double)step->stats.rss_max,
 					 outbuf, sizeof(outbuf),
 					 UNIT_KILO);
 
@@ -292,7 +292,7 @@ void print_fields(slurmdb_step_rec_t *step)
 					     (curr_inx == field_count));
 			break;
 		case PRINT_MAXVSIZE:
-			convert_num_unit((float)step->stats.vsize_max,
+			convert_num_unit((double)step->stats.vsize_max,
 					 outbuf, sizeof(outbuf),
 					 UNIT_KILO);
 
@@ -359,7 +359,7 @@ void print_fields(slurmdb_step_rec_t *step)
 			else if (step->req_cpufreq == CPU_FREQ_HIGH)
 				snprintf(outbuf, sizeof(outbuf), "High");
 			else if (!fuzzy_equal(step->req_cpufreq, NO_VAL))
-				convert_num_unit2((float)step->req_cpufreq,
+				convert_num_unit2((double)step->req_cpufreq,
 						  outbuf, sizeof(outbuf),
 						  UNIT_KILO, 1000, false);
 			field->print_routine(field,
